requirements.txt
sqlalchemy 
psycopg2-binary
python-dotenv
sentence-transformers
pymilvus
ollama
pandas